add wallet topup endpoints

+++ models/wallet.go
@@ -3,7 +3,7 @@ package models
 type Wallet struct {
 	ID       uint `gorm:"primaryKey"`
 	UserId   uint
-	Type     string `gorm:"type"` // "identified" | "unidentified"
-	Balance  int64  `gorm:"got null; default:0"`
-	Currency string `gorm:"default:'TJS'"`
+	Type     string  `gorm:"type"` // "identified" | "unidentified"
+	Balance  float64 `gorm:"got null; default:0"`
+	Currency string  `gorm:"default:'TJS'"`
 }
